HW-3: Base 생성
