+++ class14/simuelunbo/AndroidArchitectureStudy/app/src/main/java/com/example/androidarchitecturestudy/ui/BaseContract.kt
@@ -1,0 +1,17 @@
+package com.example.androidarchitecturestudy.ui
+
+import com.example.androidarchitecturestudy.data.model.Movie
+
+interface BaseContract {
+    interface View {
+        fun showLoadingBar()
+        fun hideLoadingBar()
+        fun showResultEmpty(error: String)
+        fun getMovieData(movie: List<Movie>)
+    }
+
+    interface Presenter {
+        fun saveMovieData(movie: List<Movie>)
+        fun requestLocalMovieData()
+    }
+}
