+++ class14/simuelunbo/AndroidArchitectureStudy/app/src/main/java/com/example/androidarchitecturestudy/ui/MainContract.kt
@@ -1,25 +1,17 @@
 package com.example.androidarchitecturestudy.ui
 
-import android.content.Context
-import com.example.androidarchitecturestudy.data.model.Movie
 import com.example.androidarchitecturestudy.data.model.MovieData
 
 interface MainContract {
-    interface View {
+    interface View : BaseContract.View {
         fun hideKeyboard()
-        fun showLoadingBar()
-        fun hideLoadingBar()
         fun showQueryEmpty()
-        fun showResultEmpty(error: String)
         fun showResult(movieData: MovieData)
-        fun getMovieData(movie: List<Movie>)
         fun openMovieLink(link: String)
     }
 
-    interface Presenter {
-        fun saveMovieData(movie: List<Movie>)
+    interface Presenter : BaseContract.Presenter {
         fun requestMovieInfo(title: String)
-        fun requestLocalMovieData()
         fun openMovieLink(link: String)
     }
 }
